requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
colorama==0.4.5
Django==3.2.13
django-appconf==1.0.5
django-bootstrap5==22.1
django-imagekit==4.1.0
mypy-extensions==0.4.3
pathspec==0.10.1
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
colorama==0.4.5
decorator==5.1.1
django-extensions==3.2.1
executing==1.1.0
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
pure-eval==0.2.2
Pygments==2.13.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
traitlets==5.4.0
tzdata==2022.4
wcwidth==0.2.5
